+++ src/mprXml.c
@@ -570,7 +570,7 @@ static int scanFor(MprXml *xp, char *pattern)
             /*
                 Remove the pattern from the tokBuf
              */
-            mprAdjustBufEnd(tokBuf, -strlen(pattern));
+            mprAdjustBufEnd(tokBuf, - (int) strlen(pattern));
             trimToken(xp);
             return 1;
         }
